AI/requirements.txt
# Flask, AWS, Mysql, redis, celery 등
Flask
redis
celery
boto3
load_dotenv
deemix
jsonify
mysql
pymysql
moviepy
flask_restx
ffmpeg-python